first pass over the CC models
NOTE: like the migrations, these are not the final models.
Although there are some resource specific attributes, the models
should be reviewed for hiearchry and referential integrity only.
The individual fields for each model will get updated and
thourough review as the full functionality is enabled for that
model, i.e. when apps start actually getting used, services
provisioned, etc.

Change-Id: I1394311e170ec757ba669cfbcca43ffcbd6c67e4