Integration point for Ilia version 2.
Lots was split out since the last time, and the new API is making
an appearance.  Specifically, the enumeration API is mostly in place
and all user ids now use uaa ids.  Casual users of the API are really going
to to want the inline-relations option.  That will go in next to make this
easier to use.

This commit will get a bit more breaking up and cleanup as the api is
finalized.

Change-Id: I7509a50707588b324e608e3ccd2b3e991b0ea746