merge the initial migrations into a single file
This was needed because since the migrations were being
edited after creation, some ended up referencing ones
that had later date stamps as foreign keys were added.

Sqlite was cool about this, but when this was pushed
to run with pg, pg wasn't.

During the merge, I added a VCAP::Migration helper
right in the migration file.  This should get moved
out, but let's +2/submit without that and we can
extract it later.

Change-Id: Id3da8739428f9a58b82ec2a2b5f0564f7be474b7