bump em to 1.0.0
The backported 0.12.11 was reliably segfaulting under yeti.  1.0.0 does not.

ccng makes heavier use of threads and there was a consistent segfault when
popping items off the em defer work queue and dispatching to them.  Given that
this was happening in the ruby code, the real bug was someplace out in the c
code, and somewhat removed from the actual segfault in the ruby interpreter.

We didn't see this under the 0.12.x line because our previous useses didn't
make much (if any) use of threads.

Change-Id: I6f707e205f13ef714e5cb46d3d2ebee3449ced14