refactor DeaClient.get_file_uri
Expose new DeaClient.get_file_uri_for_instance and implmented it via the now
private method DeaClient.get_file_uri.  This is in preparation for adding
DeaClient.get_file_uri_for_instance_id which is the form that vmc crashlogs
needs to use.

Note: for those that may not be familiar, an "instance" refers
to an app index from 0 to app.instances-1 and correspond to the indices
returned to the user in "vmc instances" whereas an instance_id is a guid
generated by the DEA each time it execs a droplet.

Change-Id: I762c43ce36e1f2dac129eed8ebff13e09492e3d4