Squash all migrations into a single migration
The routes for hosts were also moved to be non-nil and the empty string
is now used instead.  This is so that we can correctly perform uniquness
checks.  (The host name was being used as part of a compound unique
index, which does't work correctly at the db layer when one field is null)

Change-Id: Ifd4cf58724d5d0e06dcd029f1ecdf50489f3ee7f