Adding support for a server-side list of "supported buildpacks"
A supported buildpack has the following attributes: name, description, buildpack and support url.

Normal users using clients such as vmc will be able to read the list and provide it to the user for a better experience.
Admins will be able to maintain the list using clients such as vmc.
The list can be initialized by running the controller with the "-m' (run migrations) parameter and using a configuration file (the default being "config/supported_buildpacks.yml").

Change-Id: I75135372966356612dc7289949be33fb83db0869