Make symmetric_secret an optional config property.
The token validation logic in uaa_util.rb decides whether
to use symmetric or asymmetric key validation base on whether
this property is set. Without this change the CC won't start
unless it is set, hence it won't work with a UAA which is
configured to use an RSA key signature.