Allow Events to be queried by timestamp and type
[finishes #54656170]

* Added ability to query Events by timestamp and type of event.
* Added indices to the `timestamp` and `type` columns on the `events`
  table.

A new URL query operator has been added that allows the returning of records
that have an attribute in an array of possible values (essentially mapping to
the `IN` SQL operator). This syntax can be used as follows (pre URL escaping):

    /model?q=color IN red,blue

Further documentation about this new operator and its features and limitations
can be found in the `docs/API_Overview.md` document.