an app is staged when its droplet is uploaded
This removes the requirement that the Stager respond to an individual CC
to complete staging. For now we still send CC-specific URIs, though.

The upload endpoint currently uploads to S3 for the duration of the
request. We may convert this into a background job later.

The second request is still directed at one CC and triggers starting the
other instances, but it is not crucial to the staging having completed.

Introduced Droplet model to simplify the droplet blobstore logic. Not
fully integrated yet.