updating an app's routes sets a new version
This is so that we can send a new version in dea.update. If the message
is missed, previously the routes being registered would simply never be
updated. This allows the system to notice the receiver (the DEA) is in
an invalid state, and converge properly.

See also DEA commit b622de13d7ad6067bd969dc658f007b7ba6cc4a6.

[#57361124]