Allow apps to be scaled down or stopped when the quota is lowered.
Finishes: #52843321. User should be able to stop or scale down an app after quota is lowered.
Before the fix, when quota is lowered, memory requested always returned 0 or more, even when the requested change represented lowering of the requested memory below the memory remaining in the organization.
The above behavior caused failures when bringing the app within quota limits. The running apps could not be stopped if they were running above quota levels.