Decrease stager memory when staging/starting app
CC uses stager_advertisement to get knowledge of stager status, includes
the info like available memory. By default, these ad will be updated in
5 seconds, but when staging or starting new instance of app, the available
memory of the stager is not decerased. We need to decrease the available
memory of the stager and let CC get more accurate info of stagers

This commit will keep consistent with the previous DEA memory decrement. In
the previous commit, it just dcrease the DEA's memory while in this one, it
decreases the stager's memory