undo pull request that led to performance regression
This would make anything trying to start an app take much longer as it
did a GET of the first page of the entire S3 bucket. This would lead to
a bunch of app starts snowballing and the CC becoming unresponsive.

Revert "Added tests to confirm pull request fix"

This reverts commit e07a5124ee2c08ab33210d45881b17b198c86d21.

Revert "Fixed indentation on shared tests"

This reverts commit e68e8df4c803351ba5157f366499d36eb5176545.

Revert "Fix issue with S3 blobstore in non-standard endpoint"

This reverts commit 9d6c6f6dc3481d9ceb5bb3e9a4116c55bdb4348f.