Add 'purge' parameter to DELETE /v2/services/:guid
A delete request with the param purge=true will delete the service and
its service plans, service instances, service bindings and service plan
visibilities. The service broker will not be contacted for service
instance or binding deletes.

Due to our reliance on before_destroy and after_destroy hooks to enforce
business logic, we were not comfortable doing a Sequel::Model#delete on
each record. Instead, we set the "purged" field on the services model to
true, which replaces the service broker client with a null object. This
ensures that we do not send requests to the broker, but still executes
all before_destroy and after_destroy logic.

[#63185416]