On drain, unregister from the router and wait before shutting down
* sending USR2 signal to cloud controller will cause it to unregister from the router
* during drain, after 20 seconds, stops accepting requests.  The wait is a poor-man's
solution to increase the chance that no requests get routed to the CC during restart.
If a router doesn't receive the message, or if there is a network delay causing the
unregister to take especially long, you could still see the cloud controller receiving
requests (which would cause 502s for a user).

Signed-off-by: Kai Zhang <kaizhang@cn.ibm.com>