Allow updating services when quota is almost reached
The MaxServiceInstancePolicy assumes that all requests are to create a
service instance, and therefore checks whether the quota is less than
the quota, rather than <=. This is a problem when you're exactly at the
quota limit and try to, for example, rename a service instance. You get
a message saying you've exceeded your quota, which you haven't.