Reduce the amount of non-transactional tests
Non-transactional tests recreate the DB which is a disaster for perf and
doesn't seem to work quite right in MySQL when running in parallel. Not sure
what could be going on since each DB is theoretically isolated??? Anyway it
likes to flake out, which is of course bad. And it occasionally doesn't quite
seem to drop and recreate every table, and there's no protection since our
transactions use savepoints, so rollbacks are unpredictable.

Hopefully this change will reduce the amount of leakage and thus flakiness.

This change is the low hanging fruit. We should look to remove
`non_transactional` entirely. For that, we have [#71339198].

[#71248484]

Signed-off-by: Chris Adams <cadams@pivotallabs.com>