Complete deletion of v1 service when 404 is returned from v1 service gateway
To avoid failure on deletion of v1 service binding / instance
when v1 service gateway returns "404 Not Found".

This problem is not limited on deletion of service binding / instance.
The deletion of related app, space, and org also fails.

So this commit modifies http client of v1 service gateway to ignore
"404 Not Found" response for delete request, like client of v2 service
ignores "410 Gone" response for delete request.