NewRelic can be completely disabled
Prior to this commit, if a NewRelic license key was not provided, the
NewRelic code would still be activated in the cloud controllers and
ancillary jobs.  This code would use a default configuration that could
negatively impact memory consumption since the captured data was queued
and could not be offloaded to the NewRelic servers.

With this commit, NewRelic is enabled only when the cloud controller is
in 'development mode' or NewRelic is explicitly enabled via the
'newrelic_enabled' property in the config.  This new configuraiton
attribute is now used to guard the loading and activiation of NewRelic.

The cloud controller ctl scripts were also updated to explicitly disable
NewRelic when the deployment does not provide a license key or enable
development mode.

[#73626418]

Signed-off-by: Michael Fraenkel <fraenkel@us.ibm.com>
Signed-off-by: Matthew Sykes <matthew.sykes@gmail.com>