InstancesReporter is Diego based if app wants it
It used to fall back to the Legacy/DEA reporter when the operator
disabled diego but this would result in `cf apps` etc. incorrectly
reporting 0 instances on a DEA for an app that might have running instances on
Diego. e.g. if diego used to be on but was turned off.

[#76547300]

Signed-off-by: Peter Parente <pparent@us.ibm.com>