Remove duplicate validation in DockerPolicy
The `diego_docker` feature flag is validated by the `Backends` object
when the app is being staged. We validate it there rather than before
saving the app because the config settings could change after the app is
created.

We also removed the spec for DockerPolicy, because it now
only adds one validation, which is already exhaustively tested
in `app_spec`.

Signed-off-by: Max Brunsfeld <mbrunsfeld@pivotallabs.com>