Some more refactoring of the EventRepository
The goal here is for all methods with similar names (create_X_event) to have similar interfaces (in this case, with type, object, and request_params as the three args).

The create_event helper method is also different. It has more arguments
so that the purpose of each one is clear (we were previously overloading some arguments, like actee).