Use initializers instead of depending on Struct.new for Job classes
Because we are no longer using Struct.new, we lost some behavior that we
previously got for free.

Most importantly, Struct.new initializers allow all arguments to be
optional. We chose to make only a few of these arguments optional in our
change; only initializers that are ever actually invoked with fewer
arguments retain the originial method signature.

Another interesting change is that Struct redefines equality: two
structs are equal if all of their fields are equal. We've removed any
attempt to depend on that functionality.