External domain is no longer represented as array
Bosh template had external_domain pointing to an array, but there was
no way for a user to configure multiple external domains via the
manifest. There was also inconsistency between the config file used
for unit tests, and the actual bosh templated config. Now they are
consistent and don't unnecessarily use an array.

[#78842974]