Enlarge Service Keys Credentials
Previously the credentials column in the service_keys table was
limited to 2048 characters.  When attempting to create a service-key
against the p-rabbitmq service broker, this limit was exceeded (~2300
characters).  It's likely that this wasn't seen earlier as the
p-rabbitmq credentials payload is enormous compared to all other
service brokers.

This change marks the column as text to match the credentials column
in service_instances and to accommodate the p-rabbitmq credentials
payload.