Rename service plan visibility validation
Service plans have separate concepts for private and public, so not
being private does not mean that the plan is public. Yes, this is
confusing.

private? checks if the plan's service's broker is a private broker (has
an associated space)

public is a field that can be set on the plan to make it available to
all organizations (global)

[#97661020]

Signed-off-by: Michael Smykowski <msmykowski@pivotal.io>