TPS client returns empty array for unknown app guids
If an app is not yet running as an LRP, when we ask for instance
information from TPS, it returns a 404, which we previously handled as a
severe error and returned a 503 from the StatsController.

It is not a severe error, so we now have the same behavior as DEA apps,
which is to return an array of instances in the "DOWN" state.

[#106702544]

Signed-off-by: Dan Wendorf <dwendorf@pivotal.io>