Catch generic 'TaskCreateError' in tasks controller
The error for there being no assigned droplet for a task's parent app
seems different than the other errors we have in v3. It is not an
error with the user's request formatting (message validations) and
raising the model validation error would be confusing (task droplet
can not be nil), since the user cannot specify the droplet in their
request.

This may be the first of a new class of errors that are not
InvalidMessage or InvalidResource errors. I would rather catch a generic
error (TaskCreateError) than the specific error, so the rescue in the
controller is extensible. This way we do not need to write a controller
test for each possible action exception.

[#111472676]

Signed-off-by: Simon Leung <leungs@us.ibm.com>