Fix be_a_valid_response matcher
It added keys to the bad_key list if their values were not equal to the
expected value, but the matcher would only fail if it failed the Rspec
`match` matcher.

This meant that if you provided a regex as the expected value, the key
would get added to the bad_keys list because strings do not equal regex
values, but the matcher would not fail.

If the matcher then fails on a later key, the first key will erroniously
be in the bad keys list.

[#113011143]

Signed-off-by: Greg Cobb <gcobb@pivotal.io>