Include underlying error error in StatsUnavailable
Include the error description that caused the stats_controller of
diego to receive a StatsUnavailable exception.

Before stats_controller catched any exception, which could be caused
due network issues, misconfiguration, runtime errors, etc and raised an
error without a description. That makes difficult troubleshoot the
root cause of the issue.