Fix where private docker credentials get set
* During the process of adding support for builds, the code for adding
the docker username and password to the droplet got swallowed. This
commit adds those properties back to the droplet if they are available.

* While investigating this bug, we found that the docker_username and
docker_password fields being expected from the result of staging a
docker app actually never appear as such. So we deleted this unused
code to avoid future confusion.

[#145455375]

Signed-off-by: Jen Spinney <jennifer.spinney@suse.com>