Don't mark a build as failed if staging_complete returns an error.
* By marking the build as failed, we may overreact to some instances
where the staging_complete method returned an error, but not an error
which signifies the build itself failed. The
staging_completion_handler itself has code to mark the build as failed
when it determines a failure was reported by Diego.

* We saw an issue where Diego was calling the staging completion
callback twice for the same build, which caused the second call to mark
the build as failed, since the build was already in a final state. This
commit returns to the old behavior, which allowed the staging completion
callback to be invoked multiple times without causing a problem.

[finishes #146122097]

Signed-off-by: Jen Spinney <jennifer.spinney@suse.com>