Exec processes inside internal bin scripts
As part of a refactor to support BPM, shared functionality was moved
into a set of internal bash scripts. These scripts must `exec` the long
lived process in order to allow for proper signal handling and pidfile
management. Without the `exec`, `monit` is able to spin up multiple
instances of the processes.

[#150264913]