Emit TASK_STOPPED events when bulker fails tasks
- Previously, we manually created task start and stop events. In this
commit we add model hooks so that when a task is created, updated so
that it is in a final state (SUCCEEDED or FAILED), or destroyed, we
automatically create relevent AppUsageEvents. Keeping track of all of
the places that task state was affected proved to be less accurate for
event emitting than using model hooks.

[#150501300]

Signed-off-by: Tim Downey <tdowney@pivotal.io>