Stop failing tasks that are created while tasks_sync is running
- When task sync runs, we first fetch the list of tasks that are
currently running on BBS. Then we iterate through all of the tasks that
exist in CCDB.  We iterate through the CC tasks in batches of 500 and in
some environments this can take a non-negligible amount of time.  When
someone pushes a new task while iterating through cc tasks, that task will
not be included in the Diego tasks that we have already fetched.
- This mismatch used to cause sync to fail the task.

[#150532654]

Signed-off-by: Annie Sing <asing@pivotal.io>