Add network info to staging tasks
Network information have been recently added to the bbs task recipe.
However, they have been added to app tasks only. This patch adds
network information to staging tasks as well.

Nnote that the data will be added to the task info sent to the
BBS only when temporary_local_staging is set to True; otherwise
the stager service should be made aware of task network info as
well, which is outside the scope of this work.

PR #911
Related to #149967957

- I have viewed signed and have submitted the Contributor License Agreement
- I have made this pull request to the master branch
- I have run all the unit tests using bundle exec rake
- Network info for staging tasks have been verified with cfdot, and attaching
  to the task event stream