Correctly rehydrate Perm client in delayed jobs
The client creation sometimes adds more latency than is acceptable, so
we want to cache it by default. However, this causes a problem in that
DelayedJob has a problem serializing/deserializing these cached clients
and loggers (since connections aren't maintained across serialization).
So we provide a rehydrate method instead which is used to regenerate all
connections and loggers once deserialized in delayed jobs.

In tests, all dependencies (including the perm client) get reset each
spec. This reintroduces the latency problem and therefore makes the Perm
tests flaky because we have a low, prod-specific default timeout. We
therefore increase the acceptable timeout in tests.

We also want to parameterize the allowed timeout in production since
different consumers will have different SLOs and infrastructure
requirements.

[Finishes #151664021]

Signed-off-by: Chris Hendrix <chendrix@pivotal.io>