Stop setting #public attribute on "local" NFS blobstore files
The #public= method on local NFS blobstore Fog Files does not actually
do anything (see
https://github.com/fog/fog-local/blob/46e7a72bc19d0273e216b2986ded6e45ddd60f11/lib/fog/storage/local/models/file.rb#L69-L71)
and calling save on the file fails since the droplets are read-only on
the file system. Since #public= is a no-op we can safely remove this
code.

The tests around this were not actually testing a real scenario (they
were using an AWS Fog File that was stubbed to appear as a Local Fog
file).

[#153418863]

Signed-off-by: Lisa Cho <lcho@pivotal.io>