Fix updating encryption key for serialized field
- FieldEncryptor uses alias_method_chain to wrap field to encrypt/decrypt it
when written/read
- FieldEncryptor is used by EnvironmentVariableGroups which also have a
serializer written using alias_method_chain
- Under normal operation, they stack correctly: When hashes are written,
they are first converted to a json string and then encrypted. When the same
field is read, it is decrypted and then parsed into a hash.
- There is an edge case when moving from one encryption key to another
where fields are encrypted by directly calling the encryptor. In
this case, we need to read the field using the "field_with_encryption"
accessor because it bypasses the serializer and lets us call the
encryptor with a string instead of a hash (field_with_encryption and
field_without_serialization point to the same method due to the way the
method chain is set up).

[#150312752]

Signed-off-by: Greg Cobb <gcobb@pivotal.io>