Record a service instance delete event when space is deleted
* This change simplifies and consolidates event recording of service
instances being deleted. Previous to this change, there were several
places in the code where the event could be recorded, and we were
missing the case where a service instance is deleted via a space delete
(and the broker responds synchronously to the deprovision request).
* Also added missing unit tests for the service_instance_deprovisioner
and simplified the source code by TDDing the file.

[#152631677]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>