Add last_operation in service binding v2 resource
For synchronously created service bindings (which are the only ones
supported as of this commit), always return a last_operation object with
state 'succeeded' and operation type 'create'.

This is to make the service binding resource look uniform at the API
level when async bindings are added and consumings of the API need to
understand the last_operation key.

[#155357777]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>