Eliminate the models/v3/persistence directory
- V3 is an API version and does not make sense for models. These models
can be viewed directly or indirectly through different API versions and
should not be tied to a particular API version
- This directory existed for historical reasons. When the v3 API was
first being prototyped, it had its own, isolated models. There was a migration to
unify the models for both v2 and v3. In the modern era, this directory
no longer makes sense.