Server-side manifests: simplify error-message generation
The key here is to treat ProcessScaleMessages and
ManifestProcessScaleMessages as very different things.

In AppApplyManifest#apply I have a ManifestProcessScaleMessage
object, and call `message.process_scale_message` to get
a ProcessScaleMessage object that will be compatible with the
ProcesScale#scale method.

This commit undoes a lot of changes that were done earlier.

[#155859570]

Signed-off-by: Lisa Cho <lcho@pivotal.io>
Signed-off-by: Nick Wei <nwei@pivotal.io>