Call binmode in a separate statement
While Tempfile's binmode method, according to its implementation (https://ruby-doc.org/core-2.4.0/IO.html#method-i-binmode), returns the Tempfile object, it seems that garbage collection does not consider this a proper reference (note that the documentation also does not state that it is returning the Tempfile object). Hence, whenever garbage collection kicks in, it collects the package reference and with it, deletes the temporary file, which causes a failure further down the line when trying to work with the temp file. That's what we've seen in RelInt's Hermione env. The fact we haven't seen this anywhere else is probably because GC is undeterministic and simply wasn't executed in this specific case. Bottom line is, we used the Tempfile API wrong.

[#157250835]