Restrict service and service plan descriptions to 10000 characters
Previously, there was a database-imposed limit of 255 characters on description fields on MySQL only.

For MySQL users, this will be an increase in the possible length of the description field. For Postgres users, this will be a decrease (i.e., a new restriction).

This is therefore a breaking change for anyone who is using Postgres and has a description field in their broker catalog of larger than 10_000 characters. If such a broker exists, they will get an error the next time they update the service broker. They can fix this by reducing the size of the description to 10_000 characters or less.

[finishes #157514773]

Signed-off-by: Niki Maslarski <nikolay.maslarski@sap.com>