Stop double-requiring our v3 application_controller
Why are we making this change?
==============================

To enforce the order of before_actions, so that ProcessController's
filters do not get run before ApplicationController's filters.

It turns out that loading the before_action callback will move it from
its current place in the callback queue to the end of the queue[0][1].
We have observed that our controllers are `required`d in an
indeterminate order across various bosh-lites. We require the
application_controller explicitly and then re-require it on our
`File.expand_path`.  This demoted the application_controller's
`validate_token!` before_action, which allowed a `current_user` to be
nil in a ProcessController before_action, returning a 500 to the user
instead of a 404.

[0]: https://makandracards.com/mcsheu/39817-ordering-of-before_action-filters-in-rails-controllers
[1]: https://github.com/rails/rails/blob/e13e16f5adf3221f4a95e7d105efb2d5e56c0b53/actionpack/lib/abstract_controller/callbacks.rb#L22

[Fixes #158699580]

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>