Use rubygems version of rspec_api_documentation
We had specified the master branch of rspec_api_documentation back when
we were upgrading rspec from 2.99 to 3.0. We needed to use the master
branch because the latest release of rspec_api_documentation at the time
did not support rspec 3. We have since outgrown that need.

This commit does not change the version of rspec_api_documentation. We
still use v4.4.0. We just point to rubygems, not to GitHub.