Use the logcache client for stats instead of traffic controller
- Update proto files for logcache client and start adding tests

- Add a ruby gem which allows us to require the google/api/http in our
protoc generated files so we can stop omitting that from the generated
client stub.

- Add a TrafficController decorator, with tests, to help us convert the
logcache response to one with the same structure as the
TrafficController to avoid refactoring.

- Update CloudController.yml to add the expected logcache properties
in the api schema, allowing us to run tests. Just duplicated the
fake certs from the bbs_client into fake log_cache certs.

- Add an SSL_TARGET override in the logcache client to temporarily allow
our client to talk with logcache until the nature of their certs are
finalized.

[#159042437]

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>