Lock app, droplet, and process during "clear_process_command" migration
Added additional migration specs and manually tested some race
condition cases including updating a process's command in the middle of
the migration and deleting a process during the middle of the migration

There is still one unaccounted for "rolling-deploy" related scenario
where if a v3-push is executed after the migration has run, but before
the non-canary Cloud Controller vms have been deployed, the
"set-current-droplet" step of the v3-push will re-add the
detected/Procfile buildpack command on to the Process. We think this is
unlikely since most users will be using the regular "v2" cf push command
still. In the event this does happen, affected users will need to manually
unset the process command by setting it to "null" via the API.

[#157756113]

Signed-off-by: Eric Promislow <eric.promislow@suse.com>