Recursive app delete triggers asynchronous unbind
There might be service brokers that support only asynchronous operations.
In case a client initiates a recursive app delete
CC should act on all bindings associated with the app.

There were multiple approaches we could solve this,
but the most reasonable we come up with for now is the following:

1. Trigger asynchronous unbind (send accepts_incomplete=true to the broker)
2. Check if there is an operation in progress and raise a reasonable error to the client if there is one.

There are two v2 and v3 apps controllers that initiate that flow. Since
v2 and v3 present error messages differently (v2 can only show one
error while v3 can show multiple), we wanted the controller to decide
how the underlying action error gets presented in the API. This seemed
like a desirable separation.

* For the v2 controller we catch the error raised from the action and
re-raise nicely formated API error.

* For the v3 controller the workflow is a little bit more complex, because there
everything happens in background jobs. We have decided to introduce a
new WrappingJob that will act as an error translator. The idea is that
the we raise the same error from the action, but we handle it in our new
error_translator_job where we actualy re-raise it as a CompoundError
consisting of underlying API errors. This will result in nicely
presented list of errors that has caused the job failure.

We have added also two high level integration tests that covers the expected v2 and v3 behaviours.

[#158422116]

Signed-off-by: Georgi Lozev <georgi.lozev@sap.com>
Co-authored-by: Georgi Lozev <georgi.lozev@sap.com>
Co-authored-by: Alex Blease <ablease@pivotal.io>
Co-authored-by: Jen Spinney <jennifer.spinney@suse.com>
Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>
Co-authored-by: Luis Urraca <lurraca@pivotal.io>