Raise binding operation in progress error on service instance update
In case a client tries to update a service instance during a service
binding operation is in progress CC should raise an error.
For now we raise a general error that does not include information about
a particular binding, in v3 we can return multiple errors.

Prior to this change, broker could receive update requests while a
binding operation was in progress. This change protects the broker from
this scenario, this puts it in alignment with how CC is defensive for
concurrent broker operations. Technically this is a breaking change but one that was
considered.

[finishes #155358090]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>
Co-authored-by: Sam Gunaratne <sgunaratne@pivotal.io>