Remove duplicate errors when deleting service instance bindings
We use to propagate duplicate error massages for bindings that have
operation in progress. The duplication was caused from the fact that the
binding delete action was raising in case of operation in progress and
the instance delete action was relying also on the operation to
determine if the broker has responded asynchronously to the undbind
request.
We come up with a couple of approaches to solve this problem, however
we were not quite happy with none.
We decided to go with the approach to filter the duplicate errors
coming from the binding delete aciton since it results in the
most valuable end user massage and the least amount of components changed.

[PR]
[finishes #155358107]

Signed-off-by: Georgi Lozev <georgi.lozev@sap.com>
Co-authored-by: Georgi Lozev <georgi.lozev@sap.com>