Fix deadlock when creating/uplading a buildpack
The way we locked the buildpacks table allowed for rows to be locked by
processes without the table lock. This led to some processes owning the table
lock and wanting a row lock while another owned the row lock and wanted
the table lock. By always locking the table instead of individual rows
we avoid this deadlock. Also, relocate a lock to handle all changes
being made to the buildpacks table.

[#149010947]

Signed-off-by: Eric Promislow <eric.promislow@suse.com>