Operator can't change value of an existing encryption key
Adds an additional guardrail to prevent operators from changing the
value of an existing encryption key for a given label. This guard works
by maintaining an 'encryption_key_sentinels' table of known strings that
are encrypted with each encryption key/label pair in the deployment manifest.

At deployment time the ValidateDatabaseKeys task will run and ensure that it
can continue to decrypt the known strings in the 'encryption_key_sentinels'
table with the encryption key values that are currently configured
in the deployment manifest.

Encryption key labels that are both no-longer in use and removed from the
deployment manifest will be pruned from the 'encryption_key_sentinels'
table.

Operators should add a new encryption key label/value pair instead of
changing the value of an existing label

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>

[#160264933]

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>