Remove services_label_provider_index if present
The uniqueness constraint is set on this index, which prevents the same
service broker being registered more than once. However, this index only
exists in MySQL databases due to a difference in the way MySQL and other
DBs handle removing a column which is part of an index.

If the index has been created on multiple columns:

  :name, String
  :age, Int
  :my_sweet_index, index [:name, :age]

then if one column is dropped, MySQL will 'narrow' the index:

  :name, String
  :my_sweet_index, index [:name]

whereas e.g. Postgres will drop the index altogether. Hence the need for
the migration to first check if the index exists before attempting to
delete it.

[#154809864]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>