Display 'new_processes' on Deployments when Processes are deleted
Creates a new table, deployment_processes, for storing metadata
about the Processes that are involved during a deployment. Currently
this table only stores the "deploying_web_process" that is active during
the deployment, but it can later be extended to track the
'old_processes' or support enhanced deployments that might allow for ZDT deploying
more than a single process.

[finishes #159910933]

Signed-off-by: Chris Selzo <cselzo@pivotal.io>