Add service broker name + guid to ServiceUsageEvent
In the future, there will potentially be multiple services with the same
name. We provide the broker name and guid to help operators disambiguate
between identically-named services.

For the migration we are stuck with using "String" without a size limit
because this is how the service broker name and guid are defined in the
service_brokers table.

[finishes #159951293]
[PR]

Signed-off-by: Jen Spinney <jspinney@pivotal.io>
Signed-off-by: Henry Stanley <hstanley@pivotal.io>