Do not expect db error for duplicate service plans
It looks like in a relatively recent change as part of the ServiceManager
we've introduced a new test that expects a database validation
error is raised in case of service plan duplication. This behaviour is
not wanted anymore as part of the multi service registration work.

We should add new tests that vlidates the current implementation as
already mentioned in
https://github.com/cloudfoundry/cloud_controller_ng/commit/0629668c26ab2bfb92f156f1663f0c8de5ab3e11#diff-189fe047ec1780db79dbe8ba017a5190
commit. A PR will follow with this change, but for now just to unblock
the pipeline I'm changing only the failing unit test.
Related story https://www.pivotaltracker.com/story/show/160964169