Add tests for multiple plans and services with identical ids
With the introduction of multi service registration we could now have
services and plans with identical ids. The existing code was accidentally
working because of the way the service_manager update methods were
implemented. However there wasn't any test that covers that behaviour
and it was quite tricky to introduce such test prior merging
https://github.com/cloudfoundry/cloud_controller_ng/pull/1222.
As part of this change we add two tests that cover the desired behaviour
for updating service/plan that belongs to the
broker/service offering that is being updated. Otherwise, we might
accidentally (or maliciously) update a plan that belongs to another broker/service.
For more details have a look into the story id bellow.

[finishes #160964169]
[PR]