Remove perm integration from /v2/routes
This addresses the performance concerns for the V2 RoutesController
described in Issue #1272.

Retrofitting the v2/routes endpoints to work with the experimental
"perm" service required us to make a less-efficient database query
to fetch a list of app guids associated with the routes. Since the perm
work has been put on hiatus and the V2 endpoints are being ported over
to V3 (see: https://lists.cloudfoundry.org/g/cf-dev/topic/26659245#8305)
we can just revert this back to its previous behavior which used more
efficient joins in the Route model:

https://github.com/cloudfoundry/cloud_controller_ng/blob/ddc4ccc2ed0698bc736b0d38899b3a26b78325fe/app/models/runtime/route.rb#L152-L164

We re-introduced the 'permission enumeration' shared example to the
routes_controller_spec to ensure that the historical test coverage was
maintained.

[#162525586]

Signed-off-by: Mona Mohebbi <mmohebbi@pivotal.io>