Use TestConfig getters/setters instead of relying on it being Hash-like
Using TestConfig.config_instance.get(:configuration) instead of
TestConfig.config[:configuration] makes tests interact with the Config
object through the same means as the non-test code.

See this story for context on the great 2017 config refactor:
https://www.pivotaltracker.com/story/show/136759317

Eventually we'd like to be able to make the Config instance immutable
and this will bring us closer to that goal. See:
https://www.pivotaltracker.com/story/show/150930583

This small commit is intended to be a minimal example of updating a
test.

[#162678556]