Rename constant for '<ngx_upload_module_dummy>'
The prior name was too literal and did not convey the intent
behind this guard

This is an invalid parameter that can come through via the
nginx-upload-module during file uploads to Cloud Controller (e.g.
buildpacks, app packages, etc.). Its presence signifies that something
went awry during the upload and that some of the necessary file path
data is missing in the request from NGINX -> Cloud Controller

It comes from here:
https://github.com/shairontoledo/nginx-upload-module/blob/83fe5c02c987c25bdd0ceb6efbe3764d16984e38/ngx_http_upload_module.c#L771